GenericInterpreter

Documentation

eval :: (ArrowChoice c, ArrowFix Expr v c, ArrowEnv Text v env c, ArrowFail e c, IsString e, IsNum v c, IsClosure v env c, Join c ((v, Text), Text) v, Join c (Expr, Expr) v) => c Expr v

The Generic interpreter for PCF. It uses the IsNum and IsClosure interfaces to operate on numeric and closure values. Furthermore, it uses ArrowFix, ArrowEnv, and ArrowFail to compute the fixpoint, operate on the environment, and fail with an error message.

class Arrow c => IsNum v c | c -> v where

Interface for numeric operations

Associated Types

type Join (c :: * -> * -> *) x y :: Constraint

Methods

succ :: c v v

increments the given number value.

pred :: c v v

decrements the given number value.

zero :: c () v

creates the numeric value zero.

if_ :: Join c (x, y) z => c x z -> c y z -> c (v, (x, y)) z

class Arrow c => IsClosure v env c | c -> env, c -> v where

Interface for closures

Methods

closure :: c (Expr, env) v

creates a closure from an expression and an environment.

applyClosure :: c ((Expr, env), v) v -> c (v, v) v

applies a closure to an argument. The given continuation describes how to evaluated the body of the closure.