IntervalAnalysis

Description

k-CFA analysis for PCF where numbers are approximated by intervals.

Documentation

type Env = Map Text (Text, CallString Label) Val

data Val

Numeric values are approximated with bounded intervals, closure values are approximated with a set of abstract closures.

Constructors

NumVal IV

ClosureVal (Closure Expr Env)

Top

type Addr = (Text, CallString Label)

Addresses for this analysis are variables paired with the k-bounded call string.

evalInterval :: (?bound :: IV) => Int -> [(Text, Val)] -> State Label Expr -> Terminating (Error (Pow String) Val)

Run the abstract interpreter for the k-CFA / Interval analysis. The arguments are the maximum interval bound, the depth k of the longest call string, an environment, and the input of the computation.

newtype IntervalT c x y

Constructors

IntervalT

Fields

runIntervalT :: c x y

widening :: Widening IV -> Widening Val

type IV = Interval (InfiniteNumber Int)